% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigdriver.R
\name{sigDriver}
\alias{sigDriver}
\title{Driver discovery using mutational signature exposures}
\usage{
sigDriver(
  signature_test,
  variant_file,
  signature_file,
  covar_file,
  out_path,
  threads
)
}
\arguments{
\item{signature_test}{Name of the signature to test}

\item{variant_file}{Path to variant simple file}

\item{signature_file}{Path to tab separated signatures exposure file}

\item{covar_file}{Path to the metadata file}

\item{out_path}{Path for output}

\item{threads}{Number of threads for running}
}
\description{
Driver discovery using mutational signature exposures
}
