% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigdriver_annotate.R
\name{sigDriver_annotate}
\alias{sigDriver_annotate}
\title{sigDriver results refinement and annotation}
\usage{
sigDriver_annotate(
  signature_test,
  variant_file,
  signature_file,
  covar_file,
  out_path,
  results_file,
  annotation_gtf,
  threads
)
}
\arguments{
\item{signature_test}{Name of the signature to test}

\item{variant_file}{Path to variant simple file}

\item{signature_file}{Path to tab separated signatures exposure file}

\item{covar_file}{Path to the metadata file}

\item{out_path}{Path for output}

\item{results_file}{Path to the results file to load}

\item{annotation_gtf}{Path to the gtf file for annotation}

\item{threads}{Number of threads for running}
}
\description{
sigDriver results refinement and annotation
}
